feat: guess we're going for +Search and +Puzzle only
