+++ sv-router/src/features/puzzle/routes.ts
@@ -1,6 +1,5 @@
 import Home from '@/features/puzzle/pages/+Home.svelte';
 import Puzzle from '@/features/puzzle/pages/+Puzzle.svelte';
-import Search from '@/features/puzzle/pages/+Search.svelte';
 
 export const PUZZLE_ROUTES = {
 	HOME: '/',
@@ -10,6 +9,5 @@ export const PUZZLE_ROUTES = {
 
 export const puzzleRoutes = {
 	[PUZZLE_ROUTES.HOME]: Home,
-	[PUZZLE_ROUTES.PUZZLE]: Puzzle,
-	[PUZZLE_ROUTES.SEARCH]: Search
+	[PUZZLE_ROUTES.PUZZLE]: Puzzle
 };
