+++ sv-router/src/main.ts
@@ -1,7 +1,11 @@
 import '@/app.css';
 import { mount } from 'svelte';
 import App from '@/App.svelte';
-import { type Env, loadApplicationConfiguration, loadConfigFromVite } from '@/common/config/ApplicationConfig';
+import {
+	type Env,
+	loadApplicationConfiguration,
+	loadConfigFromVite
+} from '@/common/config/ApplicationConfig';
 
 export const env: Env = loadApplicationConfiguration(loadConfigFromVite()).getOrElse((issues) => {
 	console.error(issues.getFormattedMessages());
